requirements.txt
google-genai>=1.47.0
python-dotenv>=1.2.1
streamlit>=1.51.0
langchain>=0.3.0
langchain-google-genai>=3.0.0
langchain-core>=1.0.0
langchain-classic>=1.0.0
langchain-community>=0.4.0
pydantic>=2.0.0
pytz>=2024.1
yfinance>=0.2.0
